src/main/java/com/udemy/register/service/UserService.java
package com.udemy.register.service;

import com.udemy.register.domain.User;
import com.udemy.register.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.time.LocalDate;
import java.time.Period;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    public User register(User user) {
        var age = Period.between(user.getDateOfBirth(), LocalDate.now()).getYears();

        if (age < 18 || age > 60) {
            throw new RuntimeException("Idade não permitida");
        }

        return userRepository.save(user);
    }
}
